% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fit.R
\name{bayesian_prevalence_estimate}
\alias{bayesian_prevalence_estimate}
\title{Compute the Bayesian prevalence estimate based on the scorings}
\usage{
bayesian_prevalence_estimate(fit)
}
\arguments{
\item{fit}{The stanfit object return by BayesianFit.}
}
\value{
A numeric value of the Bayesian prevalence estimator
}
\description{
Compute the Bayesian prevalence estimate based on the scorings
}
