% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianFit.R
\name{BayesianFit}
\alias{BayesianFit}
\title{Fit the Bayesian model for Binary Replicates}
\usage{
BayesianFit(
  ni,
  si,
  prior = list(a_FP = 0.5, b_FP = 0.5, a_FN = 0.5, b_FN = 0.5, a_T = 0.5, b_T = 0.5),
  ...
)
}
\arguments{
\item{ni}{Numeric or integer vector of n_i values.}

\item{si}{Numeric or integer vector of S_i values.}

\item{prior}{A list of prior parameters for the model.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit the Bayesian model for Binary Replicates
}
