% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryReplicates-package.R
\docType{package}
\name{BinaryReplicates-package}
\alias{BinaryReplicates-package}
\alias{BinaryReplicates}
\title{The 'BinaryReplicates' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
