% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{classification_from_scoring}
\alias{classification_from_scoring}
\title{Classification based on thresholding the scorings}
\usage{
classification_from_scoring(scorings, vL, vU)
}
\arguments{
\item{scorings}{A numeric vector of the scorings}

\item{vL}{The lower threshold}

\item{vU}{The upper threshold}
}
\value{
A numeric vector of the classification (where .5 = inconclusive)
}
\description{
Classification based on thresholding the scorings
}
