% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fit.R
\name{credint}
\alias{credint}
\title{Get credible intervals for the fixed parameters of the Bayesian model}
\usage{
credint(fit, alpha = 0.05)
}
\arguments{
\item{fit}{The stanfit object return by BayesianFit.}

\item{alpha}{Error level of the credible intervals}
}
\value{
A data frame with the credible intervals
}
\description{
Get credible intervals for the fixed parameters of the Bayesian model
}
