% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammography-data.R
\docType{data}
\name{mammography-datasets}
\alias{mammography-datasets}
\alias{mammography}
\alias{observed}
\title{A mammography dataset}
\format{
A data frame \emph{mammography} with 148 rows and 3 variables
\describe{
\item{ti}{True latent state of the individual (1=positive, 0=negative)}
\item{ni}{Number of mammography tests performed for each individual}
\item{si}{Number of positive mammography tests for each individual}
}
and an 110x148 array \emph{observed} with the observed replicates, one column
for each individual.

An object of class \code{data.frame} with 148 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 110 rows and 148 columns.
}
\usage{
data(mammography, observed)

mammography

observed
}
\description{
Data from a mammography screening program.
}
\keyword{datasets}
