% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{median_based_scorings}
\alias{median_based_scorings}
\title{Compute the median-based scorings}
\usage{
median_based_scorings(ni, si)
}
\arguments{
\item{ni}{A numeric vector of the total number of replicates for each individual}

\item{si}{A numeric vector of the number of replicates equal to 1 for each individual}
}
\value{
A numeric vector of the median-based scorings
}
\description{
Compute the median-based scorings
}
