% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{most_likely_based_scorings}
\alias{most_likely_based_scorings}
\title{Compute the most likely-based scorings}
\usage{
most_likely_based_scorings(ni, si, theta, p, q)
}
\arguments{
\item{ni}{A numeric vector of the total number of replicates for each individual}

\item{si}{A numeric vector of the number of replicates equal to 1 for each individual}

\item{theta}{The probability of that T=1}

\item{p}{The false positive rate}

\item{q}{The false negative rate}
}
\value{
A numeric vector of the most likely-based scorings
}
\description{
Compute the most likely-based scorings
}
