% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{non_bayesian_scoring}
\alias{non_bayesian_scoring}
\alias{average_scoring}
\alias{median_scoring}
\alias{likelihood_scoring}
\title{Non-Bayesian scoring methods}
\usage{
average_scoring(ni, si)

median_scoring(ni, si)

likelihood_scoring(ni, si, theta, p, q)
}
\arguments{
\item{ni}{Numeric vector of \eqn{n_i}'s, the total numbers of replicates for each individual}

\item{si}{Numeric vector of \eqn{s_i}'s, the numbers of replicates equal to 1 for each individual}

\item{theta}{The probability of that \eqn{T=1}, i.e., the prevalence}

\item{p}{The false positivity rate}

\item{q}{The false negativity rate}
}
\value{
A numeric vector of the scores
}
\description{
Compute the average-, median- and likelihood-based scores
}
\examples{
data("periodontal")
Y_A <- average_scoring(periodontal$ni, periodontal$si)
Y_M <- median_scoring(periodontal$ni, periodontal$si)
theta <- mean(periodontal$ti)
cat("The prevalence in the data is ", theta, "\n")
p <- mean(Y_A[periodontal$ti == 0])
q <- mean(1-Y_A[periodontal$ti == 1])
Y_L <- likelihood_scoring(periodontal$ni, periodontal$si, theta, p, q)

}
