% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{non_bayesian_scorings}
\alias{non_bayesian_scorings}
\alias{average_based_scorings}
\alias{median_based_scorings}
\alias{most_likely_based_scorings}
\title{Non-Bayesian scorings}
\usage{
average_based_scorings(ni, si)

median_based_scorings(ni, si)

most_likely_based_scorings(ni, si, theta, p, q)
}
\arguments{
\item{ni}{A numeric vector of the total number of replicates for each individual}

\item{si}{A numeric vector of the number of replicates equal to 1 for each individual}

\item{theta}{The probability of that T=1, i.e., the prevalence}

\item{p}{The false positive rate}

\item{q}{The false negative rate}
}
\value{
A numeric vector of the scorings
}
\description{
Compute the average, median and most likely-based scorings
}
\examples{
data("periodontal")
Y_A <- average_based_scorings(periodontal$ni, periodontal$si)
Y_M <- median_based_scorings(periodontal$ni, periodontal$si)
theta <- mean(periodontal$ti)
cat("The prevalence in the data is ", theta, "\n")
p <- mean(Y_A[periodontal$ti == 0])
q <- mean(1-Y_A[periodontal$ti == 1])
Y_L <- most_likely_based_scorings(periodontal$ni, periodontal$si, theta, p, q)

}
