% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fit.R
\name{postprobT}
\alias{postprobT}
\title{Get the posterior probability that T=1 from the Bayesian model}
\usage{
postprobT(fit)
}
\arguments{
\item{fit}{The stanfit object return by BayesianFit.}
}
\value{
A vector of posterior probabilities that T=1
}
\description{
Get the posterior probability that T=1 from the Bayesian model
}
