% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_new_scores.R
\name{predict_scores}
\alias{predict_scores}
\title{Compute predictive Bayesian scores}
\usage{
predict_scores(newdata_ni, newdata_si, fit)
}
\arguments{
\item{newdata_ni}{Numeric vector of the total numbers of replicates per individuals}

\item{newdata_si}{Numeric vector of the numbers of positive replicates per individuals}

\item{fit}{The \code{stanfit} object return by \link{BayesianFit}}
}
\value{
The \code{predict_scores} function returns the predictive Bayesian scores in a numeric vector.
The predictive Bayesian scores are the posterior probabilities that the true
latent \eqn{T_i}'s are equal to 1 on new data, averaged over the posterior distribution.
}
\description{
Compute predictive Bayesian scores
}
\examples{
data("periodontal")
theta <- mean(periodontal$ti)
fit <- BayesianFit(periodontal$ni, periodontal$si, chains = 2, iter = 5000)
credint(fit)
newdata <- data.frame(ni = rep(4, 5), si = 0:4)
newdata$Y_BP <- predict_scores(newdata$ni, newdata$si, fit)
newdata
}
