% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_fit.R
\name{bayesian_computations}
\alias{bayesian_computations}
\alias{credint}
\alias{bayesian_scorings}
\alias{bayesian_prevalence_estimate}
\title{Bayesian computations}
\usage{
credint(fit, alpha = 0.05)

bayesian_scorings(fit)

bayesian_prevalence_estimate(fit)
}
\arguments{
\item{fit}{The stanfit object return by BayesianFit}

\item{alpha}{Error level of the credible intervals}
}
\value{
The \code{credint} function returns the credible interval bounds for the fixed
parameters of the Bayesian model. The default error level is 0.05.

The \code{bayesian_scoring} function returns the vector of Bayesian scorings.
These scorings are the posterior probabilities
that the true latent T\link{i} is equal to 1.

The \code{bayesian_prevalence_estimate} function returns the posterior mean of the
posterior distribution on the prevalence of T = 1.
}
\description{
Get credible intervals for the fixed parameters of the Bayesian model,
the Bayesian scorings, and the Bayesian prevalence estimate.
}
\examples{
data("periodontal")
theta <- mean(periodontal$ti)
fit <- BayesianFit(periodontal$ni, periodontal$si, chains = 2, iter = 5000)
credint(fit)
Y_B <- bayesian_scorings(fit)
T_B <- classification_from_scoring(Y_B, .4, .6)
theta_B <- bayesian_prevalence_estimate(fit)
cat("The Bayesian prevalence estimate is ", theta_B, "\n")
cat("The prevalence in the data is ", theta, "\n")

}
\seealso{
\code{\link{classification_from_scoring}}
}
