% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryReplicates-package.R
\docType{package}
\name{BinaryReplicates-package}
\alias{BinaryReplicates-package}
\alias{BinaryReplicates}
\title{BinaryReplicates: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\references{
Manuela Carenzi-Royer, Hadrien Lorenzo and Pierre Pudlo, Aix-Marseille Université.
}
\author{
\strong{Maintainer}: Pierre Pudlo \email{pierre.pudlo@univ-amu.fr} (\href{https://orcid.org/0000-0003-0995-716X}{ORCID})

Authors:
\itemize{
  \item Manuela Royez-Carenzi \email{manuela.carenzi@univ-amu.fr} (\href{https://orcid.org/0000-0003-2392-1120}{ORCID})
  \item Hadrien Lorenzo \email{hadrien.lorenzo@univ-amu.fr} (\href{https://orcid.org/0000-0002-0253-8539}{ORCID})
}

}
\keyword{internal}
