% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryReplicates-package.R
\docType{package}
\name{BinaryReplicates-package}
\alias{BinaryReplicates-package}
\alias{BinaryReplicates}
\title{BinaryReplicates: Dealing with Binary Replicates}
\description{
Statistical methods for analyzing binary replicates, which are noisy binary measurements of latent binary states. Provides scoring functions (average, median, likelihood-based, and Bayesian) to estimate the probability that an individual is in the positive state. Includes maximum a posteriori estimation via the EM algorithm and full Bayesian inference via Stan. Supports classification with inconclusive decisions and prevalence estimation.
}
\references{
Royer-Carenzi, M., Lorenzo, H., & Pudlo, P. (in press). Reconciling Binary
Replicates: Beyond the Average. \emph{Statistics in Medicine}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pierrepudlo/BinaryReplicates}
  \item Report bugs at \url{https://github.com/pierrepudlo/BinaryReplicates/issues}
}

}
\author{
\strong{Maintainer}: Pierre Pudlo \email{pierre.pudlo@univ-amu.fr} (\href{https://orcid.org/0000-0003-0995-716X}{ORCID})

Authors:
\itemize{
  \item Manuela Royez-Carenzi \email{manuela.carenzi@univ-amu.fr} (\href{https://orcid.org/0000-0003-2392-1120}{ORCID})
  \item Hadrien Lorenzo \email{hadrien.lorenzo@univ-amu.fr} (\href{https://orcid.org/0000-0002-0253-8539}{ORCID})
}

}
\keyword{internal}
