% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{classify_with_scores}
\alias{classify_with_scores}
\title{Classification based on a thresholding of the scores}
\usage{
classify_with_scores(scores, vL, vU)
}
\arguments{
\item{scores}{Numeric vector of the scores, computed with
\link{average_scoring}, \link{median_scoring} or \link{bayesian_scoring}}

\item{vL}{The lower threshold}

\item{vU}{The upper threshold}
}
\value{
A numeric vector of the classification (where .5 = inconclusive)
}
\description{
Classification based on a thresholding of the scores
}
