% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFit.R
\name{EMFit}
\alias{EMFit}
\title{Compute the \emph{Maximum-A-Posteriori} estimate with the EM algorithm}
\usage{
EMFit(
  si,
  ni,
  ti = NULL,
  prior = list(a_FP = 1, b_FP = 1, a_FN = 1, b_FN = 1),
  N_init = 20,
  maxIter = 1000,
  errorMin = 1e-07
)
}
\arguments{
\item{si}{Numeric vector of \eqn{s_i}'s, the numbers of replicates equal to 1 for each individual}

\item{ni}{Numeric vector of \eqn{n_i}'s, the total numbers of replicates for each individual}

\item{ti}{Numeric vector of \eqn{t_i}'s, the true values of the binary variable for each individual
If \code{NULL}, the EM algorithm is used to estimate the parameters. Default to \code{NULL}. See details}

\item{prior}{A list of prior parameters for the model. The prior distribution is as follows:

\itemize{
\item The false positivity rate: \eqn{p \sim \text{Beta}(a_{FP}, b_{FP})}
\item The false negativity rate: \eqn{q \sim \text{Beta}(a_{FN}, b_{FN})}
}}

\item{N_init}{The number of initializations if \code{ti} is not provided. Default to 20}

\item{maxIter}{The maximum number of iterations if EM algorithm is used. Default to 1e3}

\item{errorMin}{The minimum error computed if EM algorithm is used. Default to 1e-7}
}
\value{
A list with the following components:
\describe{
\item{score}{ The estimated values of the scores}
\item{parameters_hat}{ The estimated values of the parameters \eqn{\theta}, \eqn{p} and \eqn{q}}
}
}
\description{
Compute the \emph{Maximum-A-Posteriori} estimate with the EM algorithm
}
\details{
This function chooses its algorithm according to what is provided in the \code{ti} argument:
\describe{
\item{\code{ti} is fully provided}{The function computes the \emph{Maximum-A-Posteriori}
estimate, with an explicit formula.}
\item{\code{ti} is not provided}{The function uses the EM algorithm
to estimate the parameters.}
\item{\code{ti} is partially provided}{The function uses the EM algorithm to
estimate the parameters.}
}
}
\examples{
data("periodontal")
periodontal_ml <- EMFit(periodontal$si,periodontal$ni,periodontal$ti)
periodontal_EM <- EMFit(periodontal$si,periodontal$ni,ti = NULL)

}
\seealso{
\link{classify_with_scores}
}
