% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonBayesian_based.R
\name{prevalence_estimate}
\alias{prevalence_estimate}
\title{Compute the average,median-based prevalence estimate based on the scorings}
\usage{
prevalence_estimate(scorings)
}
\arguments{
\item{scorings}{A numeric vector of the scorings}
}
\value{
A numeric value of the prevalence estimator
}
\description{
Compute the average,median-based prevalence estimate based on the scorings
}
